requirements.txt
asyncpg
click
discord.py[speedups]
discord.py[voice]
humanfriendly
git+https://github.com/Rapptz/discord.py
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/stroupbslayen/discord-pretty-help